move AlertInsightsProvider to te-server
