+++ thirdeye-server/src/test/java/ai/startree/thirdeye/core/AlertInsightsProviderTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package ai.startree.thirdeye.alert;
+package ai.startree.thirdeye.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
